+++ active_record_has_many_split_through.gemspec
@@ -1,25 +1,25 @@
 
 lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "active_record_has_many_remotely_through/version"
+require "active_record_has_many_split_through/version"
 
 Gem::Specification.new do |spec|
-  spec.name          = "active_record_has_many_remotely_through"
-  spec.version       = ActiveRecordHasManyRemotelyThrough::VERSION
+  spec.name          = "active_record_has_many_split_through"
+  spec.version       = ActiveRecordHasManySplitThrough::VERSION
   spec.authors       = ["Nathan Herald"]
   spec.email         = ["me@nathanherald.com"]
 
   spec.summary       = %q{Support has_many :through accross seperate databases}
   spec.description   = %q{Support has_many :through accross seperate databases}
-  spec.homepage      = "https://github.com/myobie/active_record_has_many_remotely_through"
+  spec.homepage      = "https://github.com/myobie/active_record_has_many_split_through"
   spec.license       = "MIT"
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
   # to allow pushing to a single host or delete this section to allow pushing to any host.
   if spec.respond_to?(:metadata)
     spec.metadata["homepage_uri"] = spec.homepage
-    spec.metadata["source_code_uri"] = "https://github.com/myobie/active_record_has_many_remotely_through"
-    spec.metadata["changelog_uri"] = "https://github.com/myobie/active_record_has_many_remotely_through"
+    spec.metadata["source_code_uri"] = "https://github.com/myobie/active_record_has_many_split_through"
+    spec.metadata["changelog_uri"] = "https://github.com/myobie/active_record_has_many_split_through"
   else
     raise "RubyGems 2.0 or newer is required to protect against " \
       "public gem pushes."
