+++ lib/active_record/associations/builder/has_many_split_through.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module ActiveRecord::Associations::Builder # :nodoc:
-  class HasManyRemotelyThrough < HasMany #:nodoc:
+  class HasManySplitThrough < HasMany #:nodoc:
     def self.valid_options(options)
       super + [:split]
     end
