--- lib/active_record_has_many_remotely_through/version.rb
@@ -1,3 +1,0 @@
-module ActiveRecordHasManyRemotelyThrough
-  VERSION = "0.1.0"
-end
