+++ lib/active_record_has_many_split_through/version.rb
@@ -1,0 +1,3 @@
+module ActiveRecordHasManySplitThrough
+  VERSION = "0.1.0"
+end
